Factor out product into it's own component
